// src/store/actions/types.tsx
export enum LOGIN_USER {
  REQUESTED = "LOGIN_USER_REQUESTED",
  SUCCEEDED = "LOGIN_USER_SUCCEEDED",
  FAILED = "LOGIN_USER_FAILED",
}

export enum LOGOUT_USER {
  REQUESTED = "LOGOUT_USER_REQUESTED",
  SUCCEEDED = "LOGOUT_USER_SUCCEEDED",
  FAILED = "LOGOUT_USER_FAILED",
}

export enum REGISTER_USER {
  REQUESTED = "REGISTER_USER_REQUESTED",
  SUCCEEDED = "REGISTER_USER_SUCCEEDED",
  FAILED = "REGISTER_USER_FAILED",
}

export const AVATAR_CHANGED = "AVATAR_CHANGED";

export enum FETCH_USER {
  REQUESTED = "FETCH_USER_REQUESTED",
  SUCCEEDED = "FETCH_USER_SUCCEEDED",
  FAILED = "FETCH_USER_FAILED",
}

export enum DELETE_CLIENT {
  REQUESTED = "DELETE_CLIENT_REQUESTED",
  SUCCEEDED = "DELETE_CLIENT_SUCCEEDED",
  FAILED = "DELETE_CLIENT_FAILED",
}

export enum UPDATE_CLIENT {
  REQUESTED = "UPDATE_CLIENT_REQUESTED",
  SUCCEEDED = "UPDATE_CLIENT_SUCCEEDED",
  FAILED = "UPDATE_CLIENT_FAILED",
}

export enum ADD_CLIENT {
  REQUESTED = "ADD_CLIENT_REQUESTED",
  SUCCEEDED = "ADD_CLIENT_SUCCEEDED",
  FAILED = "ADD_CLIENT_FAILED",
}

export enum FETCH_POSTS {
  REQUESTED = "FETCH_POSTS_REQUESTED",
  SUCCEEDED = "FETCH_POSTS_SUCCEEDED",
  FAILED = "FETCH_POSTS_FAILED",
}

export enum ADD_POST {
  REQUESTED = "ADD_POST_REQUESTED",
  SUCCEEDED = "ADD_POST_SUCCEEDED",
  FAILED = "ADD_POST_FAILED",
}